Uploaded All

+++ src/View/GameView.java
@@ -20,25 +20,22 @@ public class GameView extends JFrame implements Serializable {
     private final int E_DOOR = 4;
     private final int E_BEGIN = 5;
     private final int E_DEST = 6;
-
     private int myPreX, myPreY = 1;
     private int myCurrentX = 1, myCurrentY = 1;
 
-    private JPanel myPanelHolder;
-    private JButton mySaveButton;
     private RoomMCView myViewMC;
     private RoomSAView myViewSA;
     private RoomTFView myViewTF;
     private String myCate = "";
-    private int myId, myGameDifficulty;
+    private int myId, myGameLevel;
     int[][] myMaze = Model.Maze.getMAZE();
     private int myRectSize = Maze.myRectSize();
-    private String fileName = "StoredData.mze";
+    private String fileName = "Assets/StoredData.txt";//mze
     private static final long serialVersionUID = 1234567890L;
 
     public GameView(int x) throws FileNotFoundException {
         if(x == 3){
-            myGameDifficulty = x;
+            myGameLevel = x;
             myMaze[5][1] = 4;
             myMaze[5][3] = 4;
             myMaze[3][8] = 4;
@@ -46,8 +43,8 @@ public class GameView extends JFrame implements Serializable {
             myMaze[8][8] = 4;
             myMaze[2][4] = 1;
         }if(x == 2){
-            myGameDifficulty = x;
-            //myMaze[5][1] = 4;
+            myGameLevel = x;
+            myMaze[5][1] = 4;
             myMaze[5][3] = 4;
             myMaze[3][8] = 4;
             //myMaze[7][11] = 4;
@@ -59,70 +56,52 @@ public class GameView extends JFrame implements Serializable {
         myViewMC = new RoomMCView();
         myViewSA = new RoomSAView();
         myViewTF = new RoomTFView();
-//        myPreX = 1;
-//        myPreY = 1;
-//        myCurrentX = 1;
-//        myCurrentY = 1;
-//        myCate = "";
-//        myId = 0;
 
-        //Saving the state
-//        try{
-//            FileOutputStream file = new FileOutputStream(fileName);
-//            ObjectOutputStream out = new ObjectOutputStream(file);
-//
-//            out.writeObject(myMaze);
-//            out.writeObject(myCurrentX);
-//            out.writeObject(myCurrentY);
-//
-//            out.close();
-//            file.close();
-//        } catch (FileNotFoundException e) {
-//            throw new RuntimeException(e);
-//        } catch (IOException e) {
-//            throw new RuntimeException(e);
-//        }
+        this.addWindowListener(new java.awt.event.WindowAdapter() {
+            @Override
+            public void windowClosing(java.awt.event.WindowEvent windowEvent) {
+                if (JOptionPane.showConfirmDialog(GameView.super.getOwner(),
+                        "Would you like to save your game?", "Save Game?",
+                        JOptionPane.YES_NO_OPTION,
+                        JOptionPane.QUESTION_MESSAGE) == JOptionPane.YES_OPTION){
+                    try {
+                        FileOutputStream file = new FileOutputStream(fileName);
+                        ObjectOutputStream out = new ObjectOutputStream(file);
+
+                        out.writeObject(myMaze);
+                        out.writeObject(myCurrentX);
+                        out.writeObject(myCurrentY);
+                        //out.writeObject(myGameLevel);
+
+                        out.close();
+                        file.close();
+                    } catch (FileNotFoundException ex) {
+                        throw new RuntimeException(ex);
+                    } catch (IOException e) {
+                        throw new RuntimeException(e);
+                    }
+                }
+            }
+        });
+
     }
 
-    public GameView(int mySavedDifficulty, int[][] mySavedMaze, int mySavedX, int mySavedY) throws FileNotFoundException {
+    public GameView(int[][] mySavedMaze, int mySavedX, int mySavedY) {
         this.myMaze = mySavedMaze;
         this.myCurrentX = mySavedX;
         this.myCurrentY = mySavedY;
-        new GameView(mySavedDifficulty);
+        prepareGUI();
+        myViewMC = new RoomMCView();
+        myViewSA = new RoomSAView();
+        myViewTF = new RoomTFView();
     }
 
     private void prepareGUI() {
-
         this.setTitle("Welcome to Trivia Maze");
-  /*      myPanelHolder = new JPanel();
-        myPanelHolder.setBackground(Color.cyan);
-        mySaveButton = new JButton("Save Game");
-        mySaveButton.setLocation(800,400);
-        myPanelHolder.add(mySaveButton);
-        this.add(myPanelHolder);
-//        JButton mySavedGame = new JButton("Saved Game");
-        mySaveButton.addActionListener(e -> {
-            try{
-                FileOutputStream file = new FileOutputStream(fileName);
-                ObjectOutputStream out = new ObjectOutputStream(file);
-
-                out.writeObject(myMaze);
-                out.writeObject(myCurrentX);
-                out.writeObject(myCurrentY);
-                out.writeObject(myGameDifficulty);
-
-                out.close();
-                file.close();
-            } catch (FileNotFoundException ex) {
-                throw new RuntimeException(ex);
-            } catch (IOException ex) {
-                throw new RuntimeException(ex);
-            }
-        });*/
-//        this.setSize(640, 480);
-        this.setExtendedState(JFrame.MAXIMIZED_BOTH);
+        this.setSize(900, 800);
         this.setLocationRelativeTo(null);
         this.setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
+        //this.setDefaultCloseOperation();
         this.setVisible(true);
 
     }
@@ -134,7 +113,6 @@ public class GameView extends JFrame implements Serializable {
 
         //draw the maze
         int check = 0;
-
 
         for (int row = 0; row < myMaze.length; row++) {
             for (int col = 0; col < myMaze[0].length; col++) {
@@ -277,6 +255,74 @@ public class GameView extends JFrame implements Serializable {
         }
     }
 
+    public void displayQuestion() {
+        Room.randomIDCategory();
+        myId = Room.getLastID();
+        myCate = Room.getLastCategory();
+
+        if (myCate.equals("MC")) {
+            myViewMC = new RoomMCView(myCate, myId);
+            myViewMC.roomShow();
+        } else if (myCate.equals("TF")) {
+            myViewTF = new RoomTFView(myCate, myId);
+            myViewTF.roomShow();
+        } else {
+            myViewSA = new RoomSAView(myCate, myId);
+            myViewSA.roomShow();
+        }
+    }
+
+    private void actionGameOver() {
+
+        try {
+            AudioInputStream ais = AudioSystem.getAudioInputStream(new File("Assets/gameover.wav"));
+            final Clip clip = AudioSystem.getClip();
+            clip.open(ais);
+            clip.setFramePosition(0);
+            clip.start();
+            String text = "Game Over!";
+            ImageIcon icon = new ImageIcon("Assets/gameover.png");
+            JOptionPane.showMessageDialog(null, text, "Check^^", JOptionPane.INFORMATION_MESSAGE, icon);
+
+        }
+        catch (LineUnavailableException ln){
+            System.out.println(ln);
+        }
+        catch (IOException e){
+            System.out.println(e);
+        }
+        catch (UnsupportedAudioFileException un){
+            System.out.println(un);
+        }
+
+
+    }
+
+    private void actionWinner() {
+
+        try {
+            AudioInputStream ais = AudioSystem.getAudioInputStream(new File("Assets/winner.wav"));
+            final Clip clip = AudioSystem.getClip();
+            clip.open(ais);
+            clip.setFramePosition(0);
+            clip.start();
+            String text = "Congratulations! You're win";
+            ImageIcon icon = new ImageIcon("Assets/championEdit1.jpeg");
+            JOptionPane.showMessageDialog(null, text, "Check^^", JOptionPane.INFORMATION_MESSAGE, icon);
+        }
+        catch (LineUnavailableException ln){
+            System.out.println(ln);
+        }
+        catch (IOException e){
+            System.out.println(e);
+        }
+        catch (UnsupportedAudioFileException un){
+            System.out.println(un);
+        }
+
+    }
+
+
     public boolean canGo(int position) {
         if(position > E_FAIL || position == E_PATH)
             return true;
@@ -354,67 +400,13 @@ public class GameView extends JFrame implements Serializable {
         return temp[row][col] == E_DEST;
     }
 
-    public void displayQuestion() {
-        Room.randomIDCategory();
-        myId = Room.getLastID();
-        myCate = Room.getLastCategory();
-
-        if (myCate.equals("MC")) {
-            myViewMC = new RoomMCView(myCate, myId);
-            myViewMC.roomShow();
-        } else if (myCate.equals("TF")) {
-            myViewTF = new RoomTFView(myCate, myId);
-            myViewTF.roomShow();
-        } else {
-            myViewSA = new RoomSAView(myCate, myId);
-            myViewSA.roomShow();
+    public void print(int[][] values) {
+        for(int row = 0; row < values.length; row++) {
+            for(int col = 0; col < values[0].length; col++) {
+                System.out.print(values[row][col] + " ");
+            }
+            System.out.println();
         }
     }
-
-    private void actionGameOver() {
-
-        try {
-            AudioInputStream ais = AudioSystem.getAudioInputStream(new File("Assets/gameover.wav"));
-            final Clip clip = AudioSystem.getClip();
-            clip.open(ais);
-            clip.setFramePosition(0);
-            clip.start();
-
-            JOptionPane.showMessageDialog(null, "Game Over!");
-        }
-        catch (LineUnavailableException ln){
-            System.out.println(ln);
-        }
-        catch (IOException e){
-            System.out.println(e);
-        }
-        catch (UnsupportedAudioFileException un){
-            System.out.println(un);
-        }
-
-
-    }
-
-    private void actionWinner() {
-
-        try {
-            AudioInputStream ais = AudioSystem.getAudioInputStream(new File("Assets/winner.wav"));
-            final Clip clip = AudioSystem.getClip();
-            clip.open(ais);
-            clip.setFramePosition(0);
-            clip.start();
-
-            JOptionPane.showMessageDialog(null, "Congratulations! You're win");
-        }
-        catch (LineUnavailableException ln){
-            System.out.println(ln);
-        }
-        catch (IOException e){
-            System.out.println(e);
-        }
-        catch (UnsupportedAudioFileException un){
-            System.out.println(un);
-        }
-
-    }
+    
 }
